msm: kgsl: Do cache flush on pages obtained using get_user_pages
Consider a scenario where user allocates anonymous memory but does
not write to it. Here the physical pages are not yet allocated. Now
when this memory is requested to be imported, a list of newly
allocated zero pages is obtained using get_user_pages(). Currently
cache flush is not done for these pages and hence GPU sees stale
data. Fix this by performing cache flush on these pages.

Change-Id: Id1e8aa20e8a9de112761732ed92f30c01088840b
Signed-off-by: Puranam V G Tejaswi <quic_pvgtejas@quicinc.com>
Signed-off-by: Sebanti Das <quic_sebadas@quicinc.com>
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>