fs: Block various userspace bloatware services from running
Unfortunately, there are several bloatware services that do not cope well
with their kernel interfaces disappearing. Furthermore, there are other
services that are nasty for different reasons and should be disabled.

But there's no way to disable them from the kernel without ungodly VFS
hacks, so that's where we're left: killing poison with poison.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[ EmanuelCN: Disable msm_irqbalance since we use SBalance ]
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>

Signed-off-by: Santhosh <san8754513543@gmail.com>